Feature/add security token

+++ app/routes/bar.js
@@ -6,6 +6,7 @@ const bar = require('../models/bar');
 
 const sequelize = require('../utils/database');
 const event = require('../models/event');
+const { checkToken } = require('../utils/tokens');
 
 /**
  * @swagger
@@ -33,8 +34,22 @@ const event = require('../models/event');
  */
 
 router.get('/', async (req, res) => {
-    let bars = await bar.findAll();
-    res.status(200).json(bars);
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
+    } else {
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
+
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 3) {
+                let bars = await bar.findAll();
+                res.status(200).json(bars);
+            }
+        }
+    }
 });
 
 /**
@@ -55,9 +70,23 @@ router.get('/', async (req, res) => {
  *              description: L'id bar saisie n'est pas connu ne base de données
  */
 router.get('/:barId', async (req, res) => {
-    let { barId } = req.params;
-    let aBar = await bar.findByPk(barId);
-    res.status(200).json(aBar);
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
+    } else {
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
+
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 1) {
+                let { barId } = req.params;
+                let aBar = await bar.findByPk(barId);
+                res.status(200).json(aBar);
+            }
+        }
+    }
 });
 
 /**
@@ -81,44 +110,58 @@ router.get('/:barId', async (req, res) => {
  *              description: Le bar existe déjà en base de données
  */
 router.post('/create', async (req, res) => {
-    if (!req.body.name) {
-        res.json('Error: name is required')
-    }
-    if (!req.body.address) {
-        res.json('Error: address is required')
-    }
-    if (!req.body.postcode) {
-        res.json('Error: postcode is required')
-    }
-    if (!req.body.city) {
-        res.json('Error: city is required')
-    }
-    if (!req.body.mail) {
-        res.json('Error: mail is required')
-    }
-    let barBDD = await bar.findOne({
-        where: {
-            name: req.body.name,
-            address: req.body.address,
-            postcode: req.body.postcode,
-            city: req.body.city,
-            mail: req.body.mail
-        }
-    })
-    if (barBDD) {
-        res.status(200).json('Bar already created');
-        console.log('Bar is already created');
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
     } else {
-        const newBar = bar.build({
-            name: req.body.name,
-            address: req.body.address,
-            postcode: req.body.postcode,
-            city: req.body.city,
-            mail: req.body.mail,
-        });
-        await newBar.save();
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
 
-        res.json("Bar is created").status(201);
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 2) {
+                if (!req.body.name) {
+                    res.json('Error: name is required')
+                }
+                if (!req.body.address) {
+                    res.json('Error: address is required')
+                }
+                if (!req.body.postcode) {
+                    res.json('Error: postcode is required')
+                }
+                if (!req.body.city) {
+                    res.json('Error: city is required')
+                }
+                if (!req.body.mail) {
+                    res.json('Error: mail is required')
+                }
+                let barBDD = await bar.findOne({
+                    where: {
+                        name: req.body.name,
+                        address: req.body.address,
+                        postcode: req.body.postcode,
+                        city: req.body.city,
+                        mail: req.body.mail
+                    }
+                })
+                if (barBDD) {
+                    res.status(200).json('Bar already created');
+                    console.log('Bar is already created');
+                } else {
+                    const newBar = bar.build({
+                        name: req.body.name,
+                        address: req.body.address,
+                        postcode: req.body.postcode,
+                        city: req.body.city,
+                        mail: req.body.mail,
+                    });
+                    await newBar.save();
+
+                    res.json("Bar is created").status(201);
+                }
+            }
+        }
     }
 });
 
@@ -145,28 +188,42 @@ router.post('/create', async (req, res) => {
  *              description: Erreurs provenant des paramètres
  */
 router.put('/update/:barId', async (req, res) => {
-    let { barId } = req.params;
-    let aBar = await bar.findByPk(barId);
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
+    } else {
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
 
-    if (req.body.name) {
-        aBar.name = req.body.name;
-    }
-    if (req.body.address) {
-        aBar.address = req.body.address;
-    }
-    if (req.body.postcode) {
-        aBar.postcode = req.body.postcode;
-    }
-    if (req.body.city) {
-        aBar.city = req.body.city;
-    }
-    if (req.body.mail) {
-        aBar.mail = req.body.mail;
-    }
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 2) {
+                let { barId } = req.params;
+                let aBar = await bar.findByPk(barId);
 
-    await aBar.save();
+                if (req.body.name) {
+                    aBar.name = req.body.name;
+                }
+                if (req.body.address) {
+                    aBar.address = req.body.address;
+                }
+                if (req.body.postcode) {
+                    aBar.postcode = req.body.postcode;
+                }
+                if (req.body.city) {
+                    aBar.city = req.body.city;
+                }
+                if (req.body.mail) {
+                    aBar.mail = req.body.mail;
+                }
 
-    res.json("Bar is updated").status(200);
+                await aBar.save();
+
+                res.json("Bar is updated").status(200);
+            }
+        }
+    }
 });
 
 /**
@@ -187,16 +244,30 @@ router.put('/update/:barId', async (req, res) => {
  *              description: Erreurs provenant des paramètres
  */
 router.delete('/delete/:barId', async (req, res) => {
-    let { barId } = req.params;
-    let aBar = await bar.findByPk(barId);
-
-    if (!aBar) {
-        res.json('Error: This barId is unknow').status(404);
-        console.log('Error: This barId is unknow');
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
     } else {
-        await aBar.destroy();
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
 
-        res.json("Bar deleted").status(200);
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 3) {
+                let { barId } = req.params;
+                let aBar = await bar.findByPk(barId);
+
+                if (!aBar) {
+                    res.json('Error: This barId is unknow').status(404);
+                    console.log('Error: This barId is unknow');
+                } else {
+                    await aBar.destroy();
+
+                    res.json("Bar deleted").status(200);
+                }
+            }
+        }
     }
 });
 
@@ -220,21 +291,35 @@ router.delete('/delete/:barId', async (req, res) => {
  *              description: Erreurs provenant des paramètres
  */
 router.get('/events/:barId', async (req, res) => {
-    let { barId } = req.params;
-
-    let aBar = await bar.findByPk(barId);
-    if (!aBar) {
-        res.json("Error: This barId is unknow in database").status(404);
-        console.log("Error: This barId is unknow in database");
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
     } else {
-        const events_from_bar = await sequelize.query(
-            "SELECT EVENT.id, EVENT.name, EVENT.description FROM EVENT LEFT JOIN BAR_EVENT ON EVENT.id = BAR_EVENT.id_event LEFT JOIN BAR ON BAR_EVENT.id_bar = BAR.id WHERE BAR.id = :id_bar",
-            {
-                replacements: { id_bar: barId },
-                type: QueryTypes.SELECT
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
+
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
+        } else {
+            if (tokenOk.role_id >= 1) {
+                let { barId } = req.params;
+
+                let aBar = await bar.findByPk(barId);
+                if (!aBar) {
+                    res.json("Error: This barId is unknow in database").status(404);
+                    console.log("Error: This barId is unknow in database");
+                } else {
+                    const events_from_bar = await sequelize.query(
+                        "SELECT EVENT.id, EVENT.name, EVENT.description FROM EVENT LEFT JOIN BAR_EVENT ON EVENT.id = BAR_EVENT.id_event LEFT JOIN BAR ON BAR_EVENT.id_bar = BAR.id WHERE BAR.id = :id_bar",
+                        {
+                            replacements: { id_bar: barId },
+                            type: QueryTypes.SELECT
+                        }
+                    );
+                    res.status(200).json(events_from_bar);
+                }
             }
-        );
-        res.status(200).json(events_from_bar);
+        }
     }
 });
 
@@ -261,39 +346,53 @@ router.get('/events/:barId', async (req, res) => {
  *              description: Erreurs provenant des paramètres
  */
 router.post('/events/:barId', async (req, res) => {
-    let { barId } = req.params;
-
-    let aBar = await bar.findByPk(barId);
-    if (!aBar) {
-        res.json("Error: This barId is unknow in database").status(404);
-        console.log("Error: This barId is unknow in database");
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
     } else {
-        let eventId = req.body.id_event;
-        let check_eventId = await event.findByPk(eventId);
-        if (!check_eventId) {
-            res.json("Error: This eventId is unknow in database").status(404);
-            console.log("Error: This eventId is unknow in database");
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
+
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
         } else {
-            const [result, metadata] = await sequelize.query(
-                "SELECT id FROM `BAR_EVENT` WHERE `id_bar` = :id_bar AND `id_event` = :id_event",
-                {
-                    replacements: { id_bar: barId, id_event: eventId },
-                    type: QueryTypes.SELECT
-                }
-            );
-            if (result) {
-                res.json("Error: This relation is already in database").status(404);
-                console.log("Error: This relation is already in database");
-            } else {
-                const bar_participe_event = await sequelize.query(
-                    "INSERT INTO `BAR_EVENT`(`id`, `id_bar`, `id_event`) VALUES (null, :id_bar, :id_event)",
-                    {
-                        replacements: { id_bar: barId, id_event: eventId },
-                        type: QueryTypes.INSERT
+            if (tokenOk.role_id >= 2) {
+                let { barId } = req.params;
+
+                let aBar = await bar.findByPk(barId);
+                if (!aBar) {
+                    res.json("Error: This barId is unknow in database").status(404);
+                    console.log("Error: This barId is unknow in database");
+                } else {
+                    let eventId = req.body.id_event;
+                    let check_eventId = await event.findByPk(eventId);
+                    if (!check_eventId) {
+                        res.json("Error: This eventId is unknow in database").status(404);
+                        console.log("Error: This eventId is unknow in database");
+                    } else {
+                        const [result, metadata] = await sequelize.query(
+                            "SELECT id FROM `BAR_EVENT` WHERE `id_bar` = :id_bar AND `id_event` = :id_event",
+                            {
+                                replacements: { id_bar: barId, id_event: eventId },
+                                type: QueryTypes.SELECT
+                            }
+                        );
+                        if (result) {
+                            res.json("Error: This relation is already in database").status(404);
+                            console.log("Error: This relation is already in database");
+                        } else {
+                            const bar_participe_event = await sequelize.query(
+                                "INSERT INTO `BAR_EVENT`(`id`, `id_bar`, `id_event`) VALUES (null, :id_bar, :id_event)",
+                                {
+                                    replacements: { id_bar: barId, id_event: eventId },
+                                    type: QueryTypes.INSERT
+                                }
+                            );
+                            res.json("Relation created").status(201);
+                            console.log("Un bar participe à un event");
+                        }
                     }
-                );
-                res.json("Relation created").status(201);
-                console.log("Un bar participe à un event");
+                }
             }
         }
     }
@@ -322,39 +421,53 @@ router.post('/events/:barId', async (req, res) => {
  *              description: Erreurs provenant des paramètres
  */
 router.delete('/events/:barId', async (req, res) => {
-    let { barId } = req.params;
-
-    let aBar = await bar.findByPk(barId);
-    if (!aBar) {
-        res.json("Error: This barId is unknow in database").status(404);
-        console.log("Error: This barId is unknow in database");
+    if (!req.headers.token) {
+        res.json('Error: You need a token').status(404);
     } else {
-        let eventId = req.body.id_event;
-        let check_eventId = await event.findByPk(eventId);
-        if (!check_eventId) {
-            res.json("Error: This eventId is unknow in database").status(404);
-            console.log("Error: This eventId is unknow in database");
+        let token = req.headers.token;
+        const tokenOk = checkToken(token);
+
+        if (!tokenOk) {
+            res.json('Error: The token is incorect').status(404);
+            console.log('Error: Wrong token');
         } else {
-            const [result, metadata] = await sequelize.query(
-                "SELECT id FROM `BAR_EVENT` WHERE `id_bar` = :id_bar AND `id_event` = :id_event",
-                {
-                    replacements: { id_bar: barId, id_event: eventId },
-                    type: QueryTypes.SELECT
-                }
-            );
-            if (!result) {
-                res.json("Error: This relation is unknow in database").status(404);
-                console.log("Error: This relation is unknow in database");
-            } else {
-                const delete_participation = await sequelize.query(
-                    "DELETE FROM `BAR_EVENT` WHERE `id` = :id",
-                    {
-                        replacements: { id: result["id"] },
-                        type: QueryTypes.DELETE
+            if (tokenOk.role_id >= 2) {
+                let { barId } = req.params;
+
+                let aBar = await bar.findByPk(barId);
+                if (!aBar) {
+                    res.json("Error: This barId is unknow in database").status(404);
+                    console.log("Error: This barId is unknow in database");
+                } else {
+                    let eventId = req.body.id_event;
+                    let check_eventId = await event.findByPk(eventId);
+                    if (!check_eventId) {
+                        res.json("Error: This eventId is unknow in database").status(404);
+                        console.log("Error: This eventId is unknow in database");
+                    } else {
+                        const [result, metadata] = await sequelize.query(
+                            "SELECT id FROM `BAR_EVENT` WHERE `id_bar` = :id_bar AND `id_event` = :id_event",
+                            {
+                                replacements: { id_bar: barId, id_event: eventId },
+                                type: QueryTypes.SELECT
+                            }
+                        );
+                        if (!result) {
+                            res.json("Error: This relation is unknow in database").status(404);
+                            console.log("Error: This relation is unknow in database");
+                        } else {
+                            const delete_participation = await sequelize.query(
+                                "DELETE FROM `BAR_EVENT` WHERE `id` = :id",
+                                {
+                                    replacements: { id: result["id"] },
+                                    type: QueryTypes.DELETE
+                                }
+                            );
+                            res.json("Relation deleted").status(200);
+                            console.log("Suppression d'une participation d'un bar");
+                        }
                     }
-                );
-                res.json("Relation deleted").status(200);
-                console.log("Suppression d'une participation d'un bar");
+                }
             }
         }
     }
